+++ cfgov/data_research/forms.py
@@ -1,8 +1,7 @@
-import django form stuff
+from django import forms
+from .models import ConferenceRegistration
 
 class ConferenceRegistrationForm(forms.ModelForm):
     class Meta:
         model = ConferenceRegistration
         fields = '__all__'
-        # fields = ['name', 'organization', 'email', 'sessions', 'foodinfo', 'accommodations']
-